Moved common helpers to providerSuite
